+++ src/main/java/com/madisp/stupid/expr/DivisionExpression.java
@@ -25,7 +25,7 @@ public class DivisionExpression implements Expression {
         if (l instanceof Double || r instanceof Double) {
             return ctx.getConverter().toDouble(l) / ctx.getConverter().toDouble(r);
         }
-        return ctx.getConverter().toInt(l) / ctx.getConverter().toInt(r);
+        return ctx.getConverter().toLong(l) / ctx.getConverter().toLong(r);
     }
 
     @Override
